changed to float

+++ src/SnapTime.cpp
@@ -54,10 +54,10 @@ TTimeCollection LoadQuery(std::string InputFile) {
 }
 
 // Only for doubles
-std::vector<std::vector<double> > InflateQuery(SnapTimeConfiguration config, TTimeCollection r,
-	std::string initTS, int duration, int granularity) {
+std::vector<std::vector<long double> > InflateQuery(SnapTimeConfiguration config, TTimeCollection r,
+	std::string initTS, long double duration, long double granularity) {
 	
-	std::vector<std::vector<double> >  result;
+	std::vector<std::vector<long double> >  result;
 	TStrV QuerySplit;
 	for (std::string split : config.SymbolicSplit) {
 		QuerySplit.Add(TStr(split.c_str()));
